+++ domain/src/commonMain/kotlin/com/blockstream/green/domain/meld/MeldDomainModule.kt
@@ -4,4 +4,5 @@ import org.koin.dsl.module
 
 val meldDomainModule = module {
     factory { GetMeldCountries(get()) }
+    factory { GetPendingMeldTransactions(get()) }
 }